properly handle down messages
If the live process monitors another process, down message is not correctly forwarded, because the channel always assumes that the parent process has terminated. As a result, the channel process will crash if a down message from a non-parent is received.

This commit switches to pattern matching which resolves the crash, and also properly forwards down messages to callback handle_info.