Expand router example and add template example
It wasn't obvious to me that I could pass session values in the router,
so I expanded the router example here to show that and linked to the
module for more options.

I also wanted to show usage from within a template, since I think it's
really cool to be able to casually drop in a live section. 😎