fix replacing mixed data types
Previously, attempting to replace a string with a map when performing
the recursive merge would result in a JavaScript error:

Uncaught TypeError: Cannot assign to read only property '0'
of string '<button type="submit">Press Me</button>'

This could be triggered with the following LiveView.

```elixir
defmodule LiveViewBugWeb.CounterLive do
  use Phoenix.LiveView

  def render(assigns) do
    ~L"""
    <form phx-submit="generate">
      <%= if @value do %>
        <%= if true do %>
          <%= @value %>
        <% end %>
      <% else %>
        <%= Phoenix.HTML.Form.submit "Press Me" %>
      <% end %>
    </form>
    """
  end

  def mount(_session, socket) do
    {:ok, assign(socket, :value, nil)}
  end

  def handle_event("generate", params, socket) do
    {:noreply, assign(socket, :value, 1)}
  end
end
```

To fix this, the type of the `targetVal` is also checked when performing
a recursive merge. If the `targetVal` is not an object, then it will
simply be replaced by the new `val.