Handle redirect + flash from initial lv mount
A live view may need to redirect with a flash message during mount - for
instance, if it's a "live show" view and determines that the current
user does not have access to the resource to be displayed.

Previously, the flash message was discarded in that case. This commit
adds support for displaying the flash message after redirecting.