Closes #255
Fixes an issue wherein the incorrect LiveView was being updated when using `live_link` or `live_redirect` with multiple
LiveViews on the page. Previously, the `root` LiveView rather the LiveView that owned a particular `link_link` was
receiving the call to `pushInternalLink`. Updated the `bindNav` `click` event listener to retrieve the owner of the
`live_link` and call `pushInternalLink` on that instead.