Refactor apprend/prepend to solve patching issues.
Previously, any appended or prepended content would not be
morphed, causing newly added nested LiveView containers or
phx-hook elements to be skipped for proper handling. The new
approach precomputes an HTML string that appears as if the
append/prepend optimizations did not exist, by comparing the
existing DOM container and adding content where it would have
been in the DOM diff to be applied. One requirement of this
new approach is all phx-update containers now require a
unique ID, as well as their immediate child nodes. An added
benefit of this approach is it allows us to get append/prepend
updates for free.