Only dispatch livelink clicks to views with a router
Since there basically can only be one main view with a router attached
it does not make sense to call handle_params on other non router views.

With this commit we mark this special view with a data attribute called
phxMain and we only push internal links
to this view instead of the root view.