use liveSocket.main.href instead of liveSocket.root.href
In #437 the concept of "main" was added and I suspect is related.  
In my own app that contains a live layout that contains a live_render 
navbar, live_links stopped working (`href on undefined`).  I tracked 
it down to this line.  I'm not sure if this change is fully sufficient or if 
there should be a null check, first try root, then try main, etc.