Fix first render back button on live redirect
Fixes https://github.com/phoenixframework/phoenix_live_view/issues/480

The issue is that in the first render when users click on the browser
back button, even though the URL changes to the previous one, the page
doesn't show the previous page content. It works fine if the user clicks
on browser next button and browser back button again, but it's not a
good User experience.

To fix that, I'm making sure the new location will be registered calling
the function `registerNewLocation(window.location)` right after pushing
state on live redirect.