Change reference clarity for `c:mount/3:params` explanation
Maybe others are not as distracted by shiny objects as I am, but I found the former version a little confusing.

It takes a second to get that `params` takes either parameters or an atom -- with great humility I might suggest that this is the cost of things being called something other than what they are.  We have changed `it` (an arguably disruptive word in this summary) to `this argument`.

As for the 'It'->'The map' conversion, at very first reading `it` could reasonably be either user-defined public information or a map. Calling out the map again clarifies the referent and introduces a non-disruptive redundancy which I think is not at all out of place to folks who want to read the LiveView documentation.