Rework `phx-submit` lifecycle description
This may need some finessing -- I've put it together according to my understanding glancing at the documentation and backing code.

The dangling `phx-submit` paragraph looked to me like it needed some love. Then I realized it was feeling lonesome so I moved it up with the other submit stuff. Before I knew it, an enumerated list of the `phx-submit` lifecycle seemed appropriate.

This change also rolls in two hyphen extractions:
'side-effects' -> 'side effects'
'use-cases'    -> 'use cases'

'side effects' is used elsewhere in the module. AFAIK 'side-effects' is an acceptable Britishism -- if it is in fact the canonically good British rendering, we could definitely go the other way on this as Elixir seems to take a Britanically influenced linguistic slant on other things.

I have never heard of 'use-cases'