Fix form UI quirks on focused updates.
Since we removed the isEqualNode optimization, a focused form
element, such as a select will behave oddly with flickering
and option jumping when the element has focus, even if we
are morphing with identical children. This fix treats all
focused elements the same as regular text inputs and avoids
patching over the user's currently focused input element.
Closes #601