Improves explanation of data-phx-error-for
I was missing this data attribute which lead to having to ask other folks on the #phoenix channel in Slack, because I am not sure that "For proper form error tag updates" was clear enough about what the result of having/not having the attribute would be. Hopefully this makes it a bit clearer for the next person.