Fix sibling component patch when component siblings lack a container. Closes #679
In cases where sibling components are at the same level, they will share
a common parent. In such cases, the parent's markup tree must be pre-computed
against the component HTML diff so we can apply a partial patch. Three
patch is built with empty siblings marked as PHX_SKIP to optimize morphing
siblings that we know will not change.