Do not forward assigns when rendering blocks
In the past, if we had this code:

    <%= if @bar do %>
      <%= @foo %>
    <% else %>
      <%= @baz %>
    <% end %>

We had to render @foo whenever @foo or @bar changed.
That's because if @bar was false and then it became
true, even though @foo did not change, we had to render
it as part of its initial render.

However, since we have changed LiveView Rendered
structs to integrate with the diff engine, the diff
engine tells us if we are in our initial render or not.
This allows us to be more conservative and compute @foo
only if @foo changed, regardless of the value in @bar.