add necessary import for installation
This import is needed, otherwise on startup, `mix phx.server` will fail with
an error stating it can't find the function `fetch_live_flash`.

In looking at the example Live View app, as seen in this commit:
https://github.com/chrismccord/phoenix_live_view_example/blob/c68368049bc39ac73fe0d66a435499f8a4415f3a/lib/demo_web/router.ex#L4
it seems necessary to add this.

This is also noted in the docs here:
https://hexdocs.pm/phoenix_live_view/Phoenix.LiveView.html#put_flash/3, however
it isn't a great first time experience to have to read all of the docs just
to get started.