Make collocated templates configurable
Collocated are place in the same directory as the corresponding LiveView
and have base names derived from the module name of the LiveView itself.

Currently, these templates must have the extension ".leex" and must be
in EEx format.

This commit allow users to specify which Phoenix LiveView engine to
use and what file extension to expect on collocated views.