Update PhoenixLiveView.send_update/2 docs (#873)
Make it more clear that `send_update/2` operates on a `Phoenix.LiveComponent`
and that it can only update components mounted on the parent LiveView.

I also made the mentioned callback links link to the actual callbacks. This
makes it more clear exactly which callback is being referenced. Additionally if
the documentation was left as it was and a `LiveView.preload/1` function was
added, then an improper link would have been created.

I also separated the paragraph that talks about the implementation from the one
that talks about the requirements since the combined paragraph would have felt
rather long. The paragraph about the implementation was reworded to make it more
clear what the subject was (the component being sent the message).

Made it more clear that the `:id` must be passed in the assigns.