Docs: Fix assign_defaults examples and a would-be warning
Keeping the `= _session` in the `mount` example because it makes sense there to clarify to the doc reader that this argument is the session. But adding an underscore since this code would result in a warning otherwise.

Skipping `= _session` in `assign_defaults` examples because it's clear by then that we're dealing with a session.