Change the error example in LiveView mount (#980)
Having just come across this myself and not know the idiomatic thing to
do, I thought it would be a good idea to change the example.

In the current example, it returns `{:error, reason}` which would raise.
This change redirects to some random "/error" page. If there is a
suggestion for a better exmaple, I am happy to change it.