Share the whole component tree instead of just root (#950)
If a component already exists on the page, LV would already
re-use its statics. However, it would do so only at the first
level.

This patch changes it so we re-use the whole component tree,
and not only the root level. It also fixes a bug where we could
point to a component that would change its structure while
rendering.

Also fixes #976.