Better handle missing components
If an event is sent to a component that was removed,
the server will response with an empty diff, so we need
to take that into account.

We also don't notify the server of pruned CIDs in case
of join patches, as any pruned CID will belong to the
previous LiveView.