work around document.activeElement being null in Internet Explorer 11 (#1006)
fixes #948

When an activeElement is removed from the page in a particular fashion,
Internet Explorer 11 does not make the body the `activeElement` and
returns `null` instead. This causes failures down in the stack when
something tries to access the element .

This patch chooses the easy approach and guards the usage against
this bug, instead of trying to solve the issue when it occurs. After
IE11 support is dropped, the workaround can be easily removed.

The comments were added to ensure this is not accidentally cleaned
up (according to spec, `activeElement` should never be `null`).