Fix nested live views failing to be discarded on remove. Closes #1050
This was morphdom related. In cases where a
keyed element is removed, the children will not
necessarily have their onBeforeNodeDiscarded callback run.
This is where we were previously leaving the child view
channel and performing cleanup. The fix is to do this
cleanup within onNodeDiscarded, which is called for all children
of the discarded keyed element.