Replace the current element with latest on undoRefs
If we try to modify a button or an input with a data-ref,
previously we would discard the update. However, the ref
reply was an empty patch, they update would never be
applied.

This patch makes it so we store the latest attempt at
patching the element and replace the element with the
latest when undoing the ref.

Since this only applies applies to inputs and buttons,
we choose to not apply the whole morphdom patch, in favor
of a simple element replace.