Fix race condition on phx-trigger-action (#1149)
Once the action is submitted, the current LiveView
terminates, which would be detected as a crash by
the client which would then try to reconnect.
If reconnecting failed (for example, because the
page is a single use token), the form submission
would be cancelled without concluding.

This commit addresses this by always disconnecting
before triggering the action. The current behaviour
can still be achieved with hooks in forms with custom
JS (which already require hooks anyway).