Introduce render_inner
Prior to this commit, parent @assigns in the inner content
wold always be bound to the changed from their last render.

This commit addresses this issue by cleaning the parent
change tracking whenever @inner_content is re-rendered
without having any changes. This means changing a component
with a @inner_content block is now more efficient.

The downside is that this is a backwards incompatible
change. As the @inner_content handling got more complex,
we introduced a `render_inner/2` function.