Fix nested live_render remounting children without ID changing
When fixing #1174 and #1277 via
953f4b64620701b35e8faeb650df9cd3c4d550d0

A bug was introduced where any re-render due to change tracking
of a nested live_render would destroy and remount the child even
when its ID did not change. This was caused by the change
of morphs getNodeKey taking the session into account, which causes
the bug because a new session is always generated if change tracking
re-encounters the live_render, but we do nothing with it if the ID
does not change. The fix is to avoid getNodeKey modifications, but
we must ensure to prepare the LiveView containers for removal on
next patch by clearing their session and ID attributse, which solves
the issue in #1174 and #1277.