add more connection info to sockets
For some tasks, I would like to store a user's IP address and user agent
as part of the audit trail. In Channels that I control, I can already do
this. This adds the capability to LiveView to get those values if they
are specified in the connect_info in the Endpoint.