Added chunk_count: option to allow_upload() with default 2
chunk_count improves upload speed on medium-high latency
connections without disrupting interactivity. When chunk_count
is 2 it means that a second chunk is being sent, while the
confirmation for the first is still outstanding. So instead of
dead waiting time from the latency we're using the latency to
transfer more in the mean time.