consume_uploaded_entry/3 preserves entries state after drop
This issue was difficult to reproduce, as the offending behaviour
would only occur immediately after dropping an external upload entry,
and even then it had the potential to leave upload config in the
expected state.

The test provided here, without the requisite changes in
LiveView.Channel fails *most* of the time, but it has the potential
to return a false positive. The trick was to trigger the faulty
behaviour concurrently.

The test, "consume_uploaded_entry maintains entries state after drop"
generates multiple upload clients, each for a single file. This gets
around the synchronous behaviour of the test UploadClient. After
starting the clients, it uploads 99% of each file. The last 1% of
each file will be uploaded concurrently since that is when the
consume function will be invoked.

Without the Channel changes, each :drop message generated by invoking
consume_uploaded_entry/3 would overwrite the previous UploadConfig,
meaning the chances were high that the last write would contain
references to entries, either in progress or completed, that had
already been consumed.