Do not use updated fingerprint trees on component update
Prior to this change, whenever an event was sent to a
component and said component added children components,
we would pass the updated component tree when trying
to reuse fingerprint trees. If the component and its
children were recursive, this could make it so the
child component referenced the new tree but told the
client it was referencing the old one. This patch
addresses this issue.

Closes #1535.