Track depth in LiveView engine vars
This was not required before because the code was
always processed in order. Take this template:

    <%= @text %>
    <%= Phoenix.LiveView.HTMLEngineTest.do_block do %><%= assigns[:not_text] %><% end %>

Using LiveEngine only, it will compile to this:

    arg0 = @text
    arg1 =
      Phoenix.LiveView.HTMLEngineTest.do_block do
        arg1 = assigns[:not_text]
        {:safe, arg1}
      end
    {:safe, [arg0, arg1]}

However, the HTMLEngine changes the others the
expressions are processed. In particular, we
process the blocks first, then the root. So we
have this:

    arg0 = @text
    arg1 =
      Phoenix.LiveView.HTMLEngineTest.do_block do
        arg0 = assigns[:not_text]
        {:safe, arg0}
      end
    {:safe, [arg0, arg1]}

If the do-block does not introduce a new scope
(i.e. it is not a macro), then arg0 inside the
block will override the external arg0. By
introducing depth, now we have:

    v0_0 = @text
    v0_1 =
      Phoenix.LiveView.HTMLEngineTest.do_block do
        v1_0 = assigns[:not_text]
        {:safe, v1_0}
      end
    {:safe, [v0_0, v0_1]}

Fully addressing the conflict.