fix form change event handling target bug
In some scenarios a form change event gets sent to the incorrect phx-target, I discovered
in our application this was happening because another script manually dispatches the
change event on an input field, inside a form.

It was interesting to discover it is somehow grabbing the correct event name, but the wrong
target. Then I dove into the Phoenix LiveView JavaScript code and discovered that while it
does get the change binding from input.form for change/input events, it then proceeds to send
the original element to JS.exec where it gets the target binding from the given element, which
is the input element, not the form.

I fixed this here, it will now assume the form is the sourceElement. I also made an attempt at
a test for this, hopefully it suffices.