Update Live layouts guide re: root layout change
While the `:layout` option to the `live` macro was partially removed
from the Live layouts guide in #1566, the sentence that is the subject
of this PR was overlooked.

If the preference would be to simplify the guide by removing any reference
to overriding the layout, then I can update this PR accordingly. However
as a newcomer to Phoenix who was just poking around to see how to set a
different root layout for a `live` route, this guide seems like a reasonable
place to discover that information.

If the PR reviewer agrees, then I will follow up with a second PR suggesting
changes to the paragraph that starts:

> "The "root" layout is shared by both "app" and "live" layouts.