Fix form recovery process with phx-change fields
If the very first field in the form has its own `phx-change` attribute
set, it breaks form recovery process on crash/re-connection. In these
circumstances front-end serializes and sends only this given field in
the payload.

This happens because in the `pushFormRecovery` we select only the first
input element in the form, and later (in `pushInput`) we check if a
given field has `change` binding set.

The proposed solution with passing down the piece of meta-information
as options (`_formRecovery: true`) might look a little bit ugly, so it's
most straightforward solution to let `pushInput` know that it's under
special circumstances.

P.S. See example of the issue in this repository:
https://github.com/cr0t/liveview-form-recovery-example