Update assigns_to_attributes/2 documentation
Currently the example in the documentation actually double renders the
`href` attribute, as it is statically declared as well as passed in the
`@extra` assign as dynamic attributes.

This merely updates it so that the component takes a `to` assign
(similar to the old fashioned `link/2` function from `phoenix_html`),
and so that the example is is "technically" correct.

I don't think the browser cares at all about the double attribute, but
when I noticed it while reading it had me scratching my head for a
moment wondering if it automatically deduped attributes or not.