Copy FileList on drop event when max_entries: 1, closes #2392 (#2393)
The files property of an input element is an immutable [FileList][1]
object that can only be replaced with another FileList. While we are
not allowed to make a new FileList object directly (there is no
public constructor), the [DataTransfer][2] object provided by the
HTML Drag and Drop API also has a files property of the same type.

For future reference, the DataTransfer API could be used to unify
more of the upload paths- the [public constructor][3] is generally
available.

For completeness, this change has one undesirable effect: if a user
drops multiple files, the text next to the file input will eventually
get out-of-sync with the number of pending uploads. This may be
solvable in a future update.

Tested on Mac OS Chrome 108, Firefox 108, and Safari 16.2.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/FileList
[2]: https://developer.mozilla.org/en-US/docs/Web/API/DataTransfer
[3]: https://caniuse.com/mdn-api_datatransfer_datatransfer