Fix auto-uploading when changing an existing input
When uploading a file using auto-upload, if a file input is already
populated, changing the input will not result in the auto-upload
triggering.

When calling `inputEl.getAttribute("data-phx-auto-upload")` the first
time, it returns a non-null value. When the input is changed, it returns
a null value.

When using `inputEl.dataset.phxAutoUpload` the value is correctly
non-null regardless of if the file input has been previously populated.

This closes #2271