Support multiple x <- y generators in :for
Previously "x <- y, a <- b" would result in a parser error. These
compound expressions are now supported but additional "for syntax" such
as the :into, :reduce and :uniq options are not supported, as well as
"filter expressions" which are instead handled by :if and appended to
the generators when compiling.