Update phoenix_component.ex
The documentation explains that the config works for components invoked by the module where the option is configured. However, at a first glance, it's very easy to miss the nuance and it might not be immediately clear how it actually works from the first time reading it. I think that adding an example that shows some context around it might make things easier for the average Phoenix user.