Allow select html tags to apply diff changes on focus
When a select input has the focus, such as when an user is changing a
values of a select, then we want to apply the changes that come from the
server. This was initially disabled specifically for FF as Brian's
comment but that might not be an issue anymore. We just installed a new
version of FF and it works ok now.

context: We started this investigation as we have a form with a regular
phx-change validation with a required constraint and the css classes on
errors were not applying on the select. We saw that the server was
sending those changes but the js part was not applying those because of
condition on the `mergeFocusedInput`.

We initially tried to just apply a new `id` to the select (based on the
select value) but that did not work nice as the `phx-no-feedback` class
was not being removed as phoenix was considering that select as a new
element.

Co-authored-by: Vinicius Negrisolo <vinicius.negrisolo@hashrocket.com>
Co-authored-by: Jack Rosa <jack.rosa@hashrocket.com>