Only extract `value` meta when element is not a form, closes #2680
`extractMeta` receives a "root element".

When this element is an input, we want to retrieve the current value.

When the element is a form, `form.key` will look up any inputs in the
form with the name "key". If an input has the name "value", then we would
retrieve the input element *as the value*, instead of getting the value
of an input.

We explicitly avoid getting the value of a form to avoid this edgecase.

At time of writing, the value attribute is valid on button, data, input,
li, meter, option, progress and param tags.