Fix `window.location = null` when connection is re-established
When a child live-view is present, and the connection drops and then is
re-established, it can result in an "error, stale" message from the LV,
which results in the page being reloaded. But `view.href` is not being
set on child liveviews when they are joined. This commit initializes
`view.href` to the value of the parent view's `href` attr to fix this
issue.