fix live components being destroyed in some circumstances (#2948)
This is a weird one:

Live components must have a single static root tag. When this
root tag does not itself contain an ID, the modifyRoot function could
accidentally create an element that contains the ID of the next inner
element instead. This could cause the component to be destroyed, as the
actual div that contained the "data-phx-component" attribute would be
removed by morphdom.