Ensure stream insert order.
Previously we were converting to a map when producing
the stream diff within __annotate__. This was encoding
a ordered json object by accident, which will break
with larger maps and provides no ordering gaurantees.
The fix is to send a list instead and lookup on the
client after building into a map.