Disable change tracking when racing a component prune. Closes #3026
We optimize components by tracking magicId's and skipping
rendering them when we know they haven't changed. However,
when racing a component prune, where the server adds a component
back to the page that the client has removed, we must reset
the change tracking of the component to disable the optimization
and render the component tree fully for the next render.

We do this by adding a new changeTracking field to the component
and resetting it to the default enabled state after we render.