fix various issues with limiting streams
https://github.com/phoenixframework/phoenix_live_view/commit/997f15cadc86703134b1755f1fa2a3808ff91b2a
introduces an issue where limits were not correctly enforced when adding
items in bulk.

It also fixes an issue with limits being applied differently on resets,
because we previously did not remove stream elements when they were being
re-added later. I think this was initially introduced to avoid issues
with live components not being correctly rendered because of the PHX_SKIP
optimization. To counteract this, we now remove all elements on reset to
force them being added in the correct order, but also store the live
component DOM nodes to correctly restore them.

References #2686.