Add reset option to assign_async/4 to clear previous results
Sometimes it is desirable to clear the results during an assign_async
- for example when performing an async search, if the term has changed
but the results are still loading, it doesn't make contextual sense
when the term doesn't match the results. In these cases it's better to
clear out the old results during loading.

This commit adds a `reset` flag to assign_async that will ignore
previous results if set to true resulting in the liveview rendering
with a default %AsyncResult{} struct.