Apply ignored attr merging consistently.
Previously we special cased data attributes only for removals
when the container is ignored. We now special case data attrs
consitently across patching. Attrs from the server (updates,
additions, removes) are synced with client for data attrs, and
we ignore all other changes to non-data attrs coming from the
server. Data attrs are useful to pass info to some js handler,
such as a WC, so we maintain the special case, but in a consistent
way.