merge value attribute as long as the property matches
We exclude the value from being merged on focused inputs, because the
user's input should always win.
We can still assign it as long as the value property is the same, though.
This prevents a situation where the updated hook is not being triggered
when an input is back in its "original state", because the attribute
was never changed.

Fixes #2163.