Update docs on dynamic inputs_for appends/prepends.
Closes #3056

Previously we incorrectly instructed the user to use a static checkbox
in the DOM for sending appends/prepends/deletes to the server.

When we introduced magic ID tracking, this broke because we were
relying on implicit reset of DOM state by the server on the
next render. The magic ID optimization skips unecessary DOM
traversals and updates, so the DOM state is not reset. We could
hack inputs to reset their state if not traversed, but the mistake
was relying on implicit state reset in the DOM that wasn't tracked
by the server in the first place. The fix is to use buttons to
send necessary changes to the server that don't require maintaining
DOM input state on either client or server.