Do not prevent submitting form when phx-click is set
To fix https://github.com/phoenixframework/phoenix_live_view/issues/3056
https://github.com/phoenixframework/phoenix_live_view/commit/fbe2226611511c89e9a61661f425e9d471cbbcb0
introduced a check that prevents submit buttons from submitting the form
when phx-click is set. This change is not necessary, because for buttons
that should not submit the form, the `type="button"` attribute should be
used instead.

Adjusts the documentation to reflect this.