Revert #2097 (#3050)
Fixes #2293.

https://github.com/phoenixframework/phoenix_live_view/pull/2097 wrapped
the join callback in an animation frame, therefore delaying the history
push until the DOM is already patched. This causes relative file paths
to be incorrectly resolved. Because #2173 adds the animation frame to
the scroll function, this should still work as expected and at least
in my manual tests it does.