Form serialization is hard, kinda
So... When trying to fix #3056, Chris changed the way the parameters
are built in serializeForm. This broke the way we were handling
select elements with the "multiple" option. I tried to fix this in #3067,
but actually broke Chris' code.
This commit refactors the serializeForm function to actually add a
temporary input element to the form to properly encode the values using
the `FormData` API. This fixes the issue with the multiple select and also
ensure that we send the value of the submitter at the correct position.

Also adds some tests that were harder to build than I'd like to admit,
because doing Ecto sort_param/drop_param stuff without Ecto is not trivial...

Thanks for reading!