fix skipped children on stream reset
When fixing stream limits I added a component restore functionality,
but this was only working with LiveComponent that were skipped and only
if they were the direct stream children. It could still happen that
nested elements inside the stream child were skipped and not restored.

This commit solves this by storing all stream child elements children
with a PHX_MAGIC_ID and restoring them later in case they were skipped.