Apply phx-no-feedback permanently, unless cleared (#3084)
* Apply phx-no-feedback permanently, unless cleared

`phx-no-feedback` is a class that is added to elements with the
`phx-feedback-for` attribute when the corresponding elements were not
interacted with yet. The problem with the current approach is that LV
of course removes the added class when the DOM is patched. Therefore we
were relying on tracking all input elements on each patch to properly
hide the feedback. This worked fine until the magicId optimization. Now,
whole sections of the DOM could be skipped on a patch, not tracking
inputs inside them and therefore not hiding feedback on the
corresponding containers.

This all was built before we had the `JS` module, which gives us a nice
way to permanently apply classes to an element, even if it is patched.
Using this functionality, we can now apply the `phx-no-feedback` class
correctly when we first render the page. This way, we don't need remove
the class on each patch, and we don't rely on tracking all inputs in
dom_patch. When we then interact with the input, or submit the form,
we show the feedback again as usual, removing the sticky class from the
feedback containers.

Fixes #3071.

* track containers, not inputs

* skip elements that are not in the DOM any more

remove no-feedback class when phx-feedback-for is removed

* add test for toggling phx-feedback-for