Always emit the last event when throttling
When throttling events, the last event was discarded, but this is not
what I would expect from a throttled event. Throttling an input should
ensure that it is only emitted at most every `delay` milliseconds, but
it should always emit the last event after the delay has passed. This
change ensures that the last event is always emitted when throttling.

This of course means that events that are emitted shortly after the last
throttled event will still be emmitted immeditately, basically starting
a new throttle window.