Ensure that events are always sent on blur, but this time correctly (#3097)
* Ensure that events are always sent on blur

Relates to: #3033
Relates to: #2318
Fixes: #3076

The old "fix" from #3033 had a fatal flaw because it would call an old
callback on blur.

This new commit adjusts the behavior to properly use the correct callback
by hooking into the cycle logic and ignoring the throttle in case of blur,
leaving the rest of the logic nearly untouched. This requires us to clear
the throttle timeout though, because otherwise we would call the callback
again after a blur, or multiple times when new events happen after blur
but before the next throttle timeout triggers.

The debounce/throttle code is not that easy to understand, but this time
I'm more confident that the solution is actually correct.
(famous last words)

* Always emit the last event when throttling

When throttling events, the last event was discarded, but this is not
what I would expect from a throttled event. Throttling an input should
ensure that it is only emitted at most every `delay` milliseconds, but
it should always emit the last event after the delay has passed. This
change ensures that the last event is always emitted when throttling.

This of course means that events that are emitted shortly after the last
throttled event will still be emmitted immeditately, basically starting
a new throttle window.