Clarify distinction between upload_errors/1 and upload_errors/2
When getting documentation served by an IDE, the first line is particularly important. I felt the difference wasn't clear at a glance as it was only the word "entry" in the middle of the sentence, hard to spot.

Reworded to make the distinction stand out and included a reverse note on `upload_errors/2` pointing back to `upload_errors/1` for symmetry.

Also noted that the output is always a list and not individual errors -- which felt particularly relevant to `upload_errors/1` as it documents only one possible error.